[Issue #33] Combine "Check for updates" with "About" dialog
