+++ DevPrompt/UI/MainWindow.xaml
@@ -94,7 +94,6 @@
             <MenuItem Header="_Help" SubmenuOpened="OnHelpMenuOpened">
                 <Separator Tag="[Plugins]" />
                 <MenuItem Header="_Report an issue" Command="{Binding ReportAnIssueCommand}" />
-                <MenuItem Header="_Check for updates" Command="{Binding CheckForUpdatesCommand}" />
                 <MenuItem Header="_About" Command="{Binding AboutCommand}" />
             </MenuItem>
         </Menu>
